test(resize): convert test to TS
